Add custom ability cooldowns
This adds a new custom command called "watch" which can be used to track
cooldowns used by other party members.  Part of knowing other people's
TP is also about knowing when their invigorate cooldown is up next, but
it's also useful to see goad and swiftcast cooldowns too.

An example usage of watch is the following:

  /e TP /watch Invigorate:inv:120

This adds a cooldown for 120 seconds for the skill named Invigorate.
It will display the cooldown next to the party members name with the
remaining seconds and the "inv" short name above it.

It is intended that users just add these cooldowns to their macro
that they use to enable the TP plugin.